Added dart tower, and squashed a whole bunch of bugs

+++ TowerDefense/src/com/example/towerdefense/FlameEnemy.java
@@ -6,7 +6,7 @@ import org.andengine.opengl.texture.region.ITextureRegion;
 public class FlameEnemy extends Enemy{
 	
 	private final static ITextureRegion pTextureRegion = TowerDefenseActivity.getSharedInstance().getFlameEnemyTextureRegion();
-	private final static float HEALTH = 3;
+	private final static int HEALTH = 3;
 	private final static float TRAVEL_SPEED = 0.5f; 
 	private final static Integer WORTH = 1;
 	private Path mPath;
